moving all STAC API calls to separate file
